Added Redshift dialect folder and an intermediate level for Redshift/PostgreSQL
