+++ parsing/src/test/PostgreSQLBaseTest.java
@@ -1,0 +1,102 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.calcite.test;
+
+import org.apache.calcite.sql.SqlDialect;
+import org.apache.calcite.sql.parser.SqlAbstractParserImpl;
+import org.apache.calcite.sql.parser.SqlParserImplFactory;
+import org.apache.calcite.sql.parser.SqlParserTest;
+import org.apache.calcite.sql.parser.SqlParserUtil;
+import org.apache.calcite.sql.parser.babel.SqlBabelParserImpl;
+
+import com.google.common.base.Throwables;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.Locale;
+import java.util.Objects;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+/**
+ * Tests the "PostgreSQL" SQL parser.
+ */
+abstract class PostgreSQLBaseTest extends SqlParserTest {
+
+  /** PostgreSQL and Redshift allow TIMESTAMP literals that contain only a
+   * date part. */
+  @Test public void testShortTimestampLiteral() {
+    sql("select timestamp '1969-07-20'")
+        .ok("SELECT TIMESTAMP '1969-07-20 00:00:00'");
+    // PostgreSQL allows the following. We should too.
+    sql("select ^timestamp '1969-07-20 1:2'^")
+        .fails("Illegal TIMESTAMP literal '1969-07-20 1:2': not in format "
+            + "'yyyy-MM-dd HH:mm:ss'"); // PostgreSQL gives 1969-07-20 01:02:00
+    sql("select ^timestamp '1969-07-20:23:'^")
+        .fails("Illegal TIMESTAMP literal '1969-07-20:23:': not in format "
+            + "'yyyy-MM-dd HH:mm:ss'"); // PostgreSQL gives 1969-07-20 23:00:00
+  }
+
+  /** In Redshift, DATE is a function. It requires special treatment in the
+   * parser because it is a reserved keyword.
+   * (Curiously, TIMESTAMP and TIME are not functions.) */
+  @Test void testDateFunction() {
+    final String expected = "SELECT DATE(`X`)\n"
+        + "FROM `T`";
+    sql("select date(x) from t").ok(expected);
+  }
+
+  /** In Redshift, PostgreSQL the DATEADD, DATEDIFF and DATE_PART functions have
+   * ordinary function syntax except that its first argument is a time unit
+   * (e.g. DAY). We must not parse that first argument as an identifier. */
+  @Test void testRedshiftFunctionsWithDateParts() {
+    final String sql = "SELECT DATEADD(day, 1, t),\n"
+        + " DATEDIFF(week, 2, t),\n"
+        + " DATE_PART(year, t) FROM mytable";
+    final String expected = "SELECT `DATEADD`(DAY, 1, `T`),"
+        + " `DATEDIFF`(WEEK, 2, `T`), `DATE_PART`(YEAR, `T`)\n"
+        + "FROM `MYTABLE`";
+
+    sql(sql).ok(expected);
+  }
+
+  /** Tests parsing PostgreSQL-style "::" cast operator. */
+  @Test void testParseInfixCast()  {
+    checkParseInfixCast("integer");
+    checkParseInfixCast("varchar");
+    checkParseInfixCast("boolean");
+    checkParseInfixCast("double");
+    checkParseInfixCast("bigint");
+
+    final String sql = "select -('12' || '.34')::VARCHAR(30)::INTEGER as x\n"
+        + "from t";
+    final String expected = ""
+        + "SELECT (- ('12' || '.34') :: VARCHAR(30) :: INTEGER) AS `X`\n"
+        + "FROM `T`";
+    sql(sql).ok(expected);
+  }
+
+  private void checkParseInfixCast(String sqlType) {
+    String sql = "SELECT x::" + sqlType + " FROM (VALUES (1, 2)) as tbl(x,y)";
+    String expected = "SELECT `X` :: " + sqlType.toUpperCase(Locale.ROOT) + "\n"
+        + "FROM (VALUES (ROW(1, 2))) AS `TBL` (`X`, `Y`)";
+    sql(sql).ok(expected);
+  }
+}
