+++ parsing/src/test/PostgreSQLParserTest.java
@@ -38,78 +38,9 @@ import static org.hamcrest.MatcherAssert.assertThat;
 /**
  * Tests the "PostgreSQL" SQL parser.
  */
-final class PostgreSQLParserTest extends SqlParserTest {
+final class PostgreSQLParserTest extends PostgreSQLBaseTest {
 
   @Override protected SqlParserImplFactory parserImplFactory() {
     return PostgreSQLParserImpl.FACTORY;
-  }
-
-  /** In Redshift, DATE is a function. It requires special treatment in the
-   * parser because it is a reserved keyword.
-   * (Curiously, TIMESTAMP and TIME are not functions.) */
-  @Test public void testDateFunction() {
-    final String expected = "SELECT `DATE`(`X`)\n"
-        + "FROM `T`";
-    sql("select date(x) from t").ok(expected);
-  }
-
-  /** PostgreSQL and Redshift allow TIMESTAMP literals that contain only a
-   * date part. */
-  @Test public void testShortTimestampLiteral() {
-    sql("select timestamp '1969-07-20'")
-        .ok("SELECT TIMESTAMP '1969-07-20 00:00:00'");
-    // PostgreSQL allows the following. We should too.
-    sql("select ^timestamp '1969-07-20 1:2'^")
-        .fails("Illegal TIMESTAMP literal '1969-07-20 1:2': not in format "
-            + "'yyyy-MM-dd HH:mm:ss'"); // PostgreSQL gives 1969-07-20 01:02:00
-    sql("select ^timestamp '1969-07-20:23:'^")
-        .fails("Illegal TIMESTAMP literal '1969-07-20:23:': not in format "
-            + "'yyyy-MM-dd HH:mm:ss'"); // PostgreSQL gives 1969-07-20 23:00:00
-  }
-
-  /** In Redshift, DATE is a function. It requires special treatment in the
-   * parser because it is a reserved keyword.
-   * (Curiously, TIMESTAMP and TIME are not functions.) */
-  @Test void testDateFunction() {
-    final String expected = "SELECT DATE(`X`)\n"
-        + "FROM `T`";
-    sql("select date(x) from t").ok(expected);
-  }
-
-  /** In Redshift, PostgreSQL the DATEADD, DATEDIFF and DATE_PART functions have
-   * ordinary function syntax except that its first argument is a time unit
-   * (e.g. DAY). We must not parse that first argument as an identifier. */
-  @Test void testRedshiftFunctionsWithDateParts() {
-    final String sql = "SELECT DATEADD(day, 1, t),\n"
-        + " DATEDIFF(week, 2, t),\n"
-        + " DATE_PART(year, t) FROM mytable";
-    final String expected = "SELECT `DATEADD`(DAY, 1, `T`),"
-        + " `DATEDIFF`(WEEK, 2, `T`), `DATE_PART`(YEAR, `T`)\n"
-        + "FROM `MYTABLE`";
-
-    sql(sql).ok(expected);
-  }
-
-  /** Tests parsing PostgreSQL-style "::" cast operator. */
-  @Test void testParseInfixCast()  {
-    checkParseInfixCast("integer");
-    checkParseInfixCast("varchar");
-    checkParseInfixCast("boolean");
-    checkParseInfixCast("double");
-    checkParseInfixCast("bigint");
-
-    final String sql = "select -('12' || '.34')::VARCHAR(30)::INTEGER as x\n"
-        + "from t";
-    final String expected = ""
-        + "SELECT (- ('12' || '.34') :: VARCHAR(30) :: INTEGER) AS `X`\n"
-        + "FROM `T`";
-    sql(sql).ok(expected);
-  }
-
-  private void checkParseInfixCast(String sqlType) {
-    String sql = "SELECT x::" + sqlType + " FROM (VALUES (1, 2)) as tbl(x,y)";
-    String expected = "SELECT `X` :: " + sqlType.toUpperCase(Locale.ROOT) + "\n"
-        + "FROM (VALUES (ROW(1, 2))) AS `TBL` (`X`, `Y`)";
-    sql(sql).ok(expected);
   }
 }
